requirements.txt
requests
numpy
fastapi
pydantic
moviepy
Pillow
arabic-reshaper
python-bidi
uvicorn
rich
